+++ src/Sql/Sql/Properties/Resources.Designer.cs
@@ -31,7 +31,7 @@ namespace Microsoft.Azure.Commands.Sql.Properties {
         [global::System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode")]
         internal Resources() {
         }
-
+        
         /// <summary>
         ///   Returns the cached ResourceManager instance used by this class.
         /// </summary>
@@ -300,6 +300,24 @@ namespace Microsoft.Azure.Commands.Sql.Properties {
         internal static string BaseConfirmActionProcessMessage {
             get {
                 return ResourceManager.GetString("BaseConfirmActionProcessMessage", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Canceling Azure Sql Managed Database Log Replay by removing &apos;{0}&apos; database..
+        /// </summary>
+        internal static string StopAzureSqlInstanceDatabaseLogReplayDescription {
+            get {
+                return ResourceManager.GetString("CancelAzureSqlInstanceDatabaseLogReplayDescription", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Are you sure you want to cancel Azure Sql Managed Database Log Replay service? By doing this specified managed database will be removed..
+        /// </summary>
+        internal static string StopAzureSqlInstanceDatabaseLogReplayWarning {
+            get {
+                return ResourceManager.GetString("CancelAzureSqlInstanceDatabaseLogReplayWarning", resourceCulture);
             }
         }
         
@@ -1247,7 +1265,7 @@ namespace Microsoft.Azure.Commands.Sql.Properties {
                 return ResourceManager.GetString("ServerNameExists", resourceCulture);
             }
         }
-
+        
         /// <summary>
         ///   Looks up a localized string similar to The server name &apos;{0}&apos; cannot be empty or null. The server name can only be made up of lowercase letters a-z, the numbers 0-9 and the hyphen. The hyphen may not lead or trail in the server name. Please fix the server name and retry. Please contact Microsoft support if the issue persists..
         /// </summary>
@@ -1416,6 +1434,15 @@ namespace Microsoft.Azure.Commands.Sql.Properties {
         internal static string StandaloneDatabaseActivityNotSupported {
             get {
                 return ResourceManager.GetString("StandaloneDatabaseActivityNotSupported", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to LastBackupName is required parameter when AutoCompleteRestore is set to true. Please explicitly provide it..
+        /// </summary>
+        internal static string StartManagedDatabaseLogReplay_LastBackupName_Warning {
+            get {
+                return ResourceManager.GetString("StartManagedDatabaseLogReplay_LastBackupName_Warning", resourceCulture);
             }
         }